add resusable wf for deploying docker
This PR adds in a modified version of existing workflows for deploying
docker containers to both DockerHub and the Github Container Registry.
Adds in additional step for optinally provisioning additional space for
builds.

Also renames the gh_release.yml workflow to indicate package release is
on Github.

Recommend to chain this as part of a deploy workflow upon a published
release (see hippunfold) or upon a pushed version tag (see scattr) to
ensure proper container metadata is extracted.