Split versioning from updates in workflows
This commit makes use of the composite actions for versioning and
commiting, splitting it from the reusable workflows. This was done
as it is difficult to predict what types of files the downstream repos
may want to update (e.g. pyproject.toml, pipeline_description.json,
etc.)

A workflow is still provided for the most generic use case: updating a
metadata file.