fix poetry caching
- Removes previously added token for deletion (this was an extra,
  unnecessary step)
- Removes the cache action that generated ${{ runner.os }}-pip-${{
  inputs.cache-id }}, this created an extra cache that contained nothing
and was not used. This means that the library stuff solely relies on the
python cache, which in the debugging logs appear to generate a new cache
if the existing one is not within the appropriate scope.
- Changed the conditionals of the installation of dependencies and
  optional library from steps.setup-poetry.outputs.success to
inputs.manager to check. Poetry should always be installed or restored
from cache, and the former conditional would set to "skip" if poetry was
restored from cache rather than success.