Add project.toml file
Create a metadata file (project.toml) that mimics relevant information
based on PEP 621 for python project metadata. Dropped the `py` in the
file name to indicate that this is not a python project - don't believe
there is a standard for GHA. This file was created with intentions to be
used for auto versioning.

Note, updating of versions within workflows and actions is still a manual
process. Part of this is to ensure the correct items are updated with
the correct context.