Checkout appropriate branch in InstallPyProject
The current workflow will always default to checking out the
`main`/`master` branch. This commit adds the reference to checkout the
PR branch, or if it does not exist, the `main`/`master` branch.

This is really only an issue if dependencies have changed in submitted
PRs, which can cause tests to fail.